feat: executive pages
